edited some grammar issues.
overall the story was pretty good. haha. I added another line to the
end. It's:
Well that was a quick story. Onwards to my next victim.

As if to say something about the narrator being a villain. Narrator
could have her personality differ depending on the choices the player
chooses.